dev-requirements.txt
asterisk-ami==0.1.6
pytest-homeassistant-custom-component
pre-commit
